requirements.txt
Django==5.1
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.2.0
django-allauth==64.1.0
daphne==4.1.2
websockets==13.0.1
requests==2.32.3
pillow==10.4.0